my_app.dartからMyHomePageを分割
ついでに名前からMyを削除
